requirements.txt
attrs
boto3
adafruit-python-shell 
adafruit-circuitpython-pm25
python-aqi